[PB-1638] refactor SharedView breadcrumbs
